+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - In Progress
 
+## 1.2.5 - 2024-07-03
+
+### Changed
+
+- Enabled analytics environment control
+- Enabled Help data to be saved to Local Storage for offline availability
+
 ## 1.2.4 - 2024-06-27
 
 ### Changed